CoordinatorFactory protocol
